[ExecuTorch][WebGPU] Promote unary op plumbing to a shared 2D-dispatch helper
Pull Request resolved: https://github.com/pytorch/executorch/pull/21153

Prep for porting the Vulkan delegate's unary activations. The elementwise-unary plumbing was file-local to the `sigmoid` op, so each new activation would duplicate it. This promotes it to a shared `runtime/ops/unary/UnaryOp.{h,cpp}` helper `add_unary_op(...)` that takes `min`/`max` params (dummy for no-param ops), mirroring the Vulkan `add_unary_op_node` (`backends/vulkan/runtime/graph/ops/impl/UnaryOp.cpp:36`), and migrates the landed `sigmoid` op onto it. The helper also adopts the 2D-fold dispatch the `add` op already uses, so unary ops on large tensors don't hit the 65535 workgroups-per-dimension cap.

Key changes:
- `runtime/ops/unary/UnaryOp.{h,cpp}` — shared `add_unary_op`: null/fp32/size guards (throw, never silent), a `UnaryParams{num_elements,min,max}` uniform mirroring Vulkan's `min_max` push constant, `compute_2d_workgroup_count` dispatch, and a resize hook that rewrites the full params (including `min`/`max`) and both dispatch dims.
- `runtime/ops/sigmoid/UnaryOp.cpp` — includes the shared header and calls `add_unary_op`; the file-local copy is removed.
- `runtime/ops/sigmoid/sigmoid.wgsl` (+ regenerated `sigmoid_wgsl.h`) — migrated to the 2D index idiom (`gid.x + gid.y*(num_workgroups.x*wg_size)`) and `wg_size=256`, matching `binary_add.wgsl`.
- `CMakeLists.txt` — add `runtime/ops/unary/UnaryOp.cpp` to `WEBGPU_SRCS`.

`sigmoid` output is unchanged: for element counts within the per-dim cap the 2D index reduces to the old 1D `gid.x`, and the map body is identical.
ghstack-source-id: 406360825
@exported-using-ghexport

Differential Revision: [D112257655](https://our.internmc.facebook.com/intern/diff/D112257655/)
